refactor(scripts): make just.config.ts part of /tasks and rename it to index
