+++ apps/pr-deploy-site/tsconfig.json
@@ -1,9 +1,10 @@
 {
-  "extends": "../../scripts/tsconfig.json",
+  "extends": "@tsconfig/node14/tsconfig.json",
   "compilerOptions": {
     "noEmit": true,
     "allowJs": true,
     "lib": ["ES2020", "DOM"]
   },
-  "include": ["pr-deploy-site.js", "just.config.ts"]
+  "include": ["pr-deploy-site.js", "just.config.ts"],
+  "files": ["../../typings/find-free-port/index.d.ts"]
 }
