+++ packages/cra-template/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../scripts/tsconfig.json",
+  "extends": "@tsconfig/node14/tsconfig.json",
   "compilerOptions": {
     "noEmit": true,
     "isolatedModules": true
