--- scripts/tsconfig.scripts.json
@@ -1,96 +1,0 @@
-{
-  "extends": "@tsconfig/node14/tsconfig.json",
-  "compilerOptions": {
-    "pretty": true,
-    "noEmit": true,
-    "allowJs": true,
-    "checkJs": true,
-    "sourceMap": true,
-    "typeRoots": ["../typings", "../node_modules/@types"]
-  },
-  "include": [],
-  "files": [],
-  "references": [
-    {
-      "path": "./api-extractor/tsconfig.json"
-    },
-    {
-      "path": "./babel/tsconfig.json"
-    },
-    {
-      "path": "./lint-staged/tsconfig.json"
-    },
-    {
-      "path": "./typescript/tsconfig.json"
-    },
-    {
-      "path": "./tasks/tsconfig.json"
-    },
-    {
-      "path": "./jest/tsconfig.json"
-    },
-    {
-      "path": "./webpack/tsconfig.json"
-    },
-    {
-      "path": "./dependency-graph-generator/tsconfig.json"
-    },
-    {
-      "path": "./triage-bot/tsconfig.json"
-    },
-    {
-      "path": "./cypress/tsconfig.json"
-    },
-    {
-      "path": "./beachball/tsconfig.json"
-    },
-    {
-      "path": "./storybook/tsconfig.json"
-    },
-    {
-      "path": "./monorepo/tsconfig.json"
-    },
-    {
-      "path": "./screener/tsconfig.json"
-    },
-    {
-      "path": "./generators/tsconfig.json"
-    },
-    {
-      "path": "./executors/tsconfig.json"
-    },
-    {
-      "path": "./github/tsconfig.json"
-    },
-    {
-      "path": "./puppeteer/tsconfig.json"
-    },
-    {
-      "path": "./gulp/tsconfig.json"
-    },
-    {
-      "path": "./dangerjs/tsconfig.json"
-    },
-    {
-      "path": "./projects-test/tsconfig.json"
-    },
-    {
-      "path": "./update-release-notes/tsconfig.json"
-    },
-    {
-      "path": "./prettier/tsconfig.json"
-    },
-    {
-      "path": "./fluentui-publish/tsconfig.json"
-    },
-    {
-      "path": "./package-manager/tsconfig.json"
-    },
-    {
-      "path": "./utils/tsconfig.json"
-    },
-    {
-      "path": "./tsconfig.json"
-    }
-  ]
-}
